Implemented check on overwrite.

+++ denest/src/logic.rs
@@ -1,5 +1,5 @@
 use commons::{
-    filter_by_extension, is_in_working_dir, read_files, FailedFileOperation, FileOperationTask,
+    filter_by_extension, read_files, FailedFileOperation, FileOperationError, FileOperationTask,
 };
 use std::{fs, io, path::PathBuf};
 pub struct Config {
@@ -26,10 +26,7 @@ impl Denest {
         } else {
             read_files(&working_dir, None)
         };
-        let filtered_files = filter_by_extension(files, &config.extensions)
-            .into_iter()
-            .filter(|file| is_in_working_dir(&working_dir, file))
-            .collect::<Vec<_>>();
+        let filtered_files = filter_by_extension(files, &config.extensions);
         let mut denest = Self {
             working_dir,
             tasks: vec![],
@@ -52,16 +49,27 @@ impl Denest {
             .collect::<Vec<FileOperationTask>>();
     }
 
-    pub fn execute(&mut self) -> (usize, usize) {
+    pub fn execute(&mut self) -> Result<(usize, usize), FileOperationError> {
+        let mut overwritten = vec![];
+        self.tasks.iter().for_each(|task| {
+            self.tasks.iter().for_each(|other_task| {
+                if task.from != other_task.from && task.to == other_task.to {
+                    overwritten.push(task.clone());
+                }
+            });
+        });
+        if !overwritten.is_empty() {
+            return Err(FileOperationError::FilesWouldOwerwrite(overwritten));
+        }
         self.tasks.iter().enumerate().for_each(|(i, task)| {
             if let Err(e) = fs::rename(&task.from, &task.to) {
                 self.failed_tasks.push((i, e))
             }
         });
-        (
+        Ok((
             self.tasks.len() - self.failed_tasks.len(),
             self.failed_tasks.len(),
-        )
+        ))
     }
 
     pub fn get_relativized_tasks(&self) -> Vec<FileOperationTask> {
